updates:
- `html` and `htmel` in separate js files
- updates to docs and README
- rebuild in dist/
- updated rollup setup
