+++ src/index.js
@@ -1,5 +1,7 @@
 var Component = require("./component.min.js")
 var tweenState = require("./tweenState.min.js")
 var emitter = require("./emitter.min.js")
+var html = require("./html.min.js")
+var htmel = require("./htmel.min.js")
 
-export { Component, tweenState, emitter }
+export { Component, tweenState, emitter, html, htmel }
